pyproject.toml
[project]
name = "blogagentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.0.10",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.1",
    "requests>=2.32.4",
    "tavily-python>=0.7.12",
    "uvicorn>=0.34.3",
    "watchdog>=6.0.0",
]
